+++ assets/javascripts/kitten/karl/headers/karl-header.js
@@ -1,14 +1,25 @@
 import React from 'react'
 import classNames from 'classnames'
 import { KisskissIcon } from 'kitten/components/icons/kisskiss-icon'
-import { InformationButton } from 'kitten/components/buttons/information-button'
+import { InformationIcon } from 'kitten/components/icons/information-icon'
+import { BurgerIcon } from 'kitten/components/icons/burger-icon'
+import { ButtonIcon } from 'kitten/components/buttons/button-con'
 
 export const KarlHeader = () => {
   return (
     <div className="karl-Menu__header">
+      <ButtonIcon className="karl-Menu__header__menu" { ...props }>
+        <BurgerIcon className="k-ButtonIcon__svg" />
+      </ButtonIcon>
+
       <KisskissIcon className="karl-Menu__header__logo" />
+
       <h3 className="karl-Menu__header__title">Styleguide</h3>
-      <InformationButton className="karl-Menu__header__about" />
+
+      <ButtonIcon className="karl-Menu__header__about" { ...props }>
+        <InformationIcon className="k-ButtonIcon__svg" />
+      </ButtonIcon>
+
     </div>
   )
 }