+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Add `Separator` component.
+- Feature: Add `Separator` component.
 - Fix: Install new grid on each components.
 - Fix: Delete example descriptions.
 - Fix: Redesign new subtitles and add border-bottom.