+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Remove `button-heart-icon-words` to `button-with-icon-words`.
+- Fix: Replace `KarlButtonIconHeartRight` and `KarlButtonIconHeartLeft` examples
+  with `KarlButtonIconWordsRight` and `KarlButtonIconWordsLeft`.
 - Fix: Propagate `className` prop to `Grid` and `GridCol` components.
 - Fix: Remove `error` label on Checkbox and RadioButton guideline example.
 - Feature: Add `Separator` component.