+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Update `hexa` colors and layout.
+- Fix: Replace `KarlButtonIconHeartRight` and `KarlButtonIconHeartLeft` examples
+  with `KarlButtonIconWordsRight` and `KarlButtonIconWordsLeft`.
+- Feature: Add `normalLineHeight` prop to `Paragraph` component.
+- Fix: Propagate `className` prop to `Grid` and `GridCol` components.
 - Fix: Remove `error` label on Checkbox and RadioButton guideline example.
 - Feature: Add `Separator` component.
 - Fix: Install new grid on each components.